refactor(convert): extract inline_root module with try_convert entry
Phase 1 of fulgur-odgo. Move inline-root handling from convert/mod.rs
into convert/inline_root.rs:

- pub(super) fn try_convert(doc, node_id, ctx, depth) -> Option<...>:
  wraps the dispatcher's `if node.flags.is_inline_root() { ... }`
  branch, including the paragraph + inline-pseudo + block-pseudo path.
  Each `return Box::new(...)` becomes `return Some(Box::new(...))`.
  Fall-through (inline root with no text and no inline pseudo images)
  drops to the bottom None and the dispatcher continues to its
  childless/container path.
- pub(super): extract_paragraph, recalculate_paragraph_line_boxes,
  metrics_from_line, resolve_enclosing_anchor.
- private: convert_inline_box_node (only called by extract_paragraph).

Sibling submodules updated to call via inline_root::<name>:
- list_item.rs: extract_paragraph, recalculate_paragraph_line_boxes,
  metrics_from_line.
- list_marker.rs: recalculate_paragraph_line_boxes.
- pseudo.rs: resolve_enclosing_anchor.
mod.rs LinkCache::lookup also updated for resolve_enclosing_anchor,
and the 6 extract_paragraph call sites in mod tests are updated to
super::inline_root::extract_paragraph.

No behavioral change. VRT byte-identical against committed goldens.
